Convert Annotation sub-components to TS
