+++ ui/src/alerts/components/SearchBar.tsx
@@ -12,7 +12,7 @@ interface State {
 
 @ErrorHandling
 class SearchBar extends PureComponent<Props, State> {
-  constructor(props) {
+  constructor(props: Props) {
     super(props)
 
     this.state = {
